EventsList Update
Renamed Screen1 to EventsList. Added a list of events to EventsList (using dummy date from TestData/testData.json) and made the events touchable (they currently display an alert, but will go to the events details page when it is implemented).

Also changed settings icon to a refresh icon since we have an app drawer now.